Rename module state-machine -> mpc.
And move the websocket client into the mpc module for now, later we may
move it to a package.
